move code to shared project
