+++ AssemblyLoadDebuggerShared/ToolWindow/AssemblyLoadDebuggerControl.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using System.Windows.Controls;
+using ToolWindow;
 
 namespace AssemblyLoadDebugger
 {
