+++ AssemblyLoadDebuggerShared/ToolWindow/AssemblyLoadDebuggerToolWindow.cs
@@ -1,9 +1,9 @@
-﻿namespace AssemblyLoadDebugger
-{
-    using System;
-    using System.Runtime.InteropServices;
-    using Microsoft.VisualStudio.Shell;
+﻿using System.Runtime.InteropServices;
+using AssemblyLoadDebugger;
+using Microsoft.VisualStudio.Shell;
 
+namespace ToolWindow
+{
     [Guid("71604717-b15d-4d1e-9fb4-03f5122e8858")]
     public class AssemblyLoadDebuggerToolWindow : ToolWindowPane
     {
