+++ AssemblyLoadDebuggerShared/VSCommandTable.cs
