+++ AssemblyLoadDebuggerShared/VSCommandTable.vsct
