+++ src/AssemblyLoadDebugger.csproj
@@ -60,27 +60,10 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="ToolWindow\Settings.cs" />
-    <Compile Include="VSCommandTable.cs">
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-      <DependentUpon>VSCommandTable.vsct</DependentUpon>
-    </Compile>
     <Compile Include="source.extension.cs">
       <DependentUpon>source.extension.vsixmanifest</DependentUpon>
     </Compile>
-    <Compile Include="ToolWindow\ActionCommand.cs" />
-    <Compile Include="ToolWindow\AssemblyLoadDebuggerControlViewModel.cs" />
-    <Compile Include="Commands\OpenToolWindowCommand.cs" />
-    <Compile Include="ToolWindow\AssemblyLoadDebuggerControl.xaml.cs">
-      <DependentUpon>AssemblyLoadDebuggerControl.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="VSPackage.cs" />
-    <Compile Include="ToolWindow\AssemblyLoadDebuggerToolWindow.cs" />
-    <Compile Include="ToolWindow\AssetCache.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ToolWindow\SharedVsTheme.cs" />
-    <Compile Include="ToolWindow\WpfUtil.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="..\LICENSE">
@@ -95,23 +78,9 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <Content Include="Resources\Icon.png">
-      <IncludeInVSIX>true</IncludeInVSIX>
-    </Content>
     <Content Include="source.extension.ico">
       <DependentUpon>source.extension.vsixmanifest</DependentUpon>
     </Content>
-    <VSCTCompile Include="VSCommandTable.vsct">
-      <ResourceName>Menus.ctmenu</ResourceName>
-      <Generator>VsctGenerator</Generator>
-      <LastGenOutput>VSCommandTable.cs</LastGenOutput>
-    </VSCTCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <Page Include="ToolWindow\AssemblyLoadDebuggerControl.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
   </ItemGroup>
   <ItemGroup>
     <Reference Include="EnvDTE, Version=8.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
@@ -178,6 +147,7 @@
     </PackageReference>
     <PackageReference Include="Microsoft.VisualStudio.CommandBars" Version="8.0.0.0" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="..\AssemblyLoadDebuggerShared\AssemblyLoadDebuggerShared.projitems" Label="Shared" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(VSToolsPath)\VSSDK\Microsoft.VsSDK.targets" Condition="'$(VSToolsPath)' != ''" />
